feat(ci): add github actions support (#7)
This change adds support for GitHub actions for every push to master and on every PR intended to be
merged to master. For the moment, it will execute cargo fmt, cargo clippy and cargo tests. It will
fail now as we have lots of format and clippy warnings but as we haven't set the branch policies, it
won't be a stopper for now. A couple of unnecessary `mut` keywords have been removed from the tests. This change also adds support for cargo-make.